pyproject.toml
[project]
name = "Motivational-Meme-Generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.3.3",
    "pandas>=2.0.3",
    "pillow>=12.0.0",
    "python-docx>=1.2.0",
]
